fix(profile): add auth and profile checks to profile page
